requirements.txt
python-telegram-bot==20.7
PyPDF2==3.0.1
reportlab==4.0.7